% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Count the number of the missing values in a vector}
\usage{
count_na(x)
}
\arguments{
\item{x}{Vector. A vector containing missing values}
}
\value{
the number of the missing values in the target vector
}
\description{
Count the number of the missing values in a vector
}
\examples{
\dontrun{
x <- c(1, 2, 3, NA, 4, 5, NA, 6, NA, 7)
count_na(x)
library(data.table)
dt <- data.table(
x = c(1, 2, 3, NA, 4, 5, NA, 6, NA, 7),
y = c(1, NA, 3, NA, 4, 5, NA, 6, NA, 7)
)
dt[, lapply(.SD, count_na)]
}
}
