% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_linear_slope.R
\name{get_linear_slope}
\alias{get_linear_slope}
\title{Get the linear slope when passing the significance test}
\usage{
get_linear_slope(obj, sig, fill = NA)
}
\arguments{
\item{obj}{A \code{lm} object.}

\item{sig}{Numeric. The significance level for testing. If not asigned, pass the test.}

\item{fill}{Numeric. The value used when not passing the significance test}
}
\value{
If the model is significant return the linear slope, else \code{NA}.
}
\description{
Get the linear slope when passing the significance test
}
\examples{
\dontrun{
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm_obj <- lm(weight ~ group)
lm_obj2 <- lm(weight ~ group - 1)
get_linear_slope(lm_obj, 0.05)
get_linear_slope(lm_obj2, 0.05)
}
}
