% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_objects.R
\name{ls_objects}
\alias{ls_objects}
\alias{lsos}
\title{Improved list of objects}
\usage{
ls_objects(pos = 1, pattern, order.by, decreasing = FALSE,
  head = FALSE, n = 5)

lsos(..., n = 10)
}
\arguments{
\item{pos}{an alternative argument to name for specifying the environment as a position in the search list.
Mostly there for back compatibility.}

\item{pattern}{an optional regular expression. Only names matching pattern are returned.
glob2rx can be used to convert wildcard patterns to regular expressions.}

\item{order.by}{the column to sort by}

\item{decreasing}{logical. Should the sort order be increasing or decreasing?}

\item{head}{logical. Should return first n parts of the results?}

\item{n}{a single integer. If positive or zero, size for the resulting object: number of elements for a vector (including lists),
rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.}
}
\value{
a data.frame
}
\description{
Improved list of objects

shorthand version of ls_objects
}
